+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added support for Azure Cognitive Search vector store (#7469)
 - Support delete in supabase (#6951)
 
+### Bug Fixes
+- Catch validation errors for structured responses (#7523)
+
 ## [0.8.20] - 2023-09-04
 
 ### New Features