+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ### New Features
 - Added support for Azure Cognitive Search vector store (#7469)
+- Support delete in supabase (#6951)
+
+### Bug Fixes / Nits
+- Default to user-configurable top-k in `VectorIndexAutoRetriever` (#7556)
+- Catch validation errors for structured responses (#7523)
 
 ## [0.8.20] - 2023-09-04